updating interfaces used within the header
